feat: queryClient to make requests and functions to calculate read time and format date added
the custom useFetch hook was replaced by the React Query tool due to the request limit that the api
has for developers, the persistQueryClient plugin was added and the data from the requests is being
cached, and stored in the localStorage.
the formatDate function was added, it receives the date in
the form that is supplied by the api and returns "Today", "Yesterday" or the date in the format
dd/mm/yyyy.
the function queryClient was added, it receives the content of the article, in the form
that is supplied by the api and returns a string containing the amount of minutes needed to read the
article considering a reading speed of 200 words per minute.